Changed write_color to convert color values to integer

+++ color.py
@@ -16,7 +16,7 @@ def write_color(pixel_color, samples_per_pixel):
     g *= scale
     b *= scale
 
-    r = 256 * clamp(r, 0.0, 0.999)
-    g = 256 * clamp(g, 0.0, 0.999)
-    b = 256 * clamp(b, 0.0, 0.999)
+    r = int(256 * clamp(r, 0.0, 0.999))
+    g = int(256 * clamp(g, 0.0, 0.999))
+    b = int(256 * clamp(b, 0.0, 0.999))
     print(r, ' ', g, ' ', b, '\n')
